% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arcticreport_cache}
\alias{arcticreport_cache}
\title{Show list of cached arcticreport files}
\usage{
arcticreport_cache()
}
\value{
(list) A list of files
}
\description{
This function returns a list of files cached for the package.
}
