% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_citations.R
\name{count_citations}
\alias{count_citations}
\title{Count number data citations}
\usage{
count_citations(from = as.POSIXct("2010-01-01"), to = as.POSIXct(Sys.Date()))
}
\arguments{
\item{from}{Start date to count over (chatacter or POSIXct)}

\item{to}{End date to count over (character of POSIXct)}
}
\value{
Number of citations in given time period
}
\description{
Count number data citations
}
