% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_creators.R
\name{count_creators}
\alias{count_creators}
\title{Count number of unique creators}
\usage{
count_creators(
  objects,
  from = as.POSIXct("1899-01-01"),
  to = as.POSIXct(Sys.Date())
)
}
\arguments{
\item{objects}{(data.frame) Table obtained from \code{query_version_chains}}

\item{from}{Start date to count over (chatacter or POSIXct)}

\item{to}{End date to count over (character of POSIXct)}
}
\value{
Number of creators
}
\description{
Count number of unique creators
}
