% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_data_objects.R
\name{count_data_objects}
\alias{count_data_objects}
\title{Count number of new data objects}
\usage{
count_data_objects(
  objects,
  from = as.POSIXct("1899-01-01"),
  to = as.POSIXct(Sys.Date())
)
}
\arguments{
\item{objects}{(data.frame) Table obtained from \code{query_objects}}

\item{from}{Start date to count over (chatacter or POSIXct)}

\item{to}{End date to count over (character of POSIXct)}
}
\value{
Number of new data objects
}
\description{
Count number of new data objects
}
