% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_datasets_theme.R
\name{count_datasets_theme}
\alias{count_datasets_theme}
\title{Count number of datasets by theme}
\usage{
count_datasets_theme(
  from = as.POSIXct("2000-01-01"),
  to = as.POSIXct(Sys.Date()),
  theme_data
)
}
\arguments{
\item{from}{Start date to count over (chatacter or POSIXct)}

\item{to}{End date to count over (character of POSIXct)}

\item{theme_data}{Data on themes read from googlesheet (data.frame)}
}
\value{
data.frame of number of datasets by theme
}
\description{
Count number of datasets by theme
}
