% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_downloads.R
\name{count_downloads}
\alias{count_downloads}
\title{Count number data downloads}
\usage{
count_downloads(from = as.POSIXct("2010-01-01"), to = as.POSIXct(Sys.Date()))
}
\arguments{
\item{from}{Start date to count over (chatacter or POSIXct)}

\item{to}{End date to count over (character of POSIXct)}
}
\value{
Number of downloads in given time period
}
\description{
Count number data downloads
}
