% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_new_and_changed_datasets.R
\name{count_new_and_changed_datasets}
\alias{count_new_and_changed_datasets}
\title{Count number of new and updated datasets}
\usage{
count_new_and_changed_datasets(
  objects,
  from = as.POSIXct("1899-01-01"),
  to = as.POSIXct(Sys.Date())
)
}
\arguments{
\item{objects}{(data.frame) Table obtained from \code{query_objects}}

\item{from}{Start date to count over (chatacter or POSIXct)}

\item{to}{End date to count over (character of POSIXct)}
}
\value{
Number of new and changed datasets
}
\description{
Count number of new and updated datasets
}
