% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_support_interactions.R
\name{count_support_interactions}
\alias{count_support_interactions}
\title{Count number of support interactions}
\usage{
count_support_interactions(
  from = as.POSIXct("2010-01-01"),
  to = as.POSIXct(Sys.Date())
)
}
\arguments{
\item{from}{Start date to count over (chatacter or POSIXct)}

\item{to}{End date to count over (character of POSIXct)}
}
\value{
Number of support interactions
}
\description{
Count number of support interactions
}
