% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_unique_accessors.R
\name{count_unique_accessors}
\alias{count_unique_accessors}
\title{Count number of people accessing data}
\usage{
count_unique_accessors(
  path,
  from = as.POSIXct("1899-01-01"),
  to = as.POSIXct(Sys.Date())
)
}
\arguments{
\item{path}{Path to file with downloads information from Metacat}

\item{from}{Start date to count over (chatacter or POSIXct)}

\item{to}{End date to count over (character of POSIXct)}
}
\value{
Number of unique people accessing data
}
\description{
Count number of people accessing data
}
\note{
In the postgres database, use query: COPY (
SELECT * FROM access_log WHERE
date_logged > '2021-02-01 00:00' AND
date_logged < '2021-04-30 23:59' AND
lower(event) = 'read'
ORDER BY date_logged ASC
)
TO '/tmp/access_log.csv' WITH CSV HEADER;
}
