% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_volume.R
\name{count_volume}
\alias{count_volume}
\title{Count total volume}
\usage{
count_volume(
  objects,
  from = as.POSIXct("1899-01-01"),
  to = as.POSIXct(Sys.Date())
)
}
\arguments{
\item{objects}{(data.frame) Table obtained from \code{query_objects}}

\item{from}{Start date to count over (chatacter or POSIXct)}

\item{to}{End date to count over (character of POSIXct)}
}
\value{
Volume in TB of data center
}
\description{
Count total volume
}
