% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme_adc.R
\name{plot_theme_adc}
\alias{plot_theme_adc}
\title{Theme function for plots}
\usage{
plot_theme_adc(font_size = 12)
}
\arguments{
\item{font_size}{(integer) Base size for font}
}
\value{
ggplot theme object
}
\description{
Theme function for plots
}
