% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_filesys_objects.R
\name{query_filesys_objects}
\alias{query_filesys_objects}
\title{Return key information on objects stored on the filesystem outside of Metacat}
\usage{
query_filesys_objects()
}
\value{
(data.frame) Result of filesystem stat
}
\description{
This function returns the identifier, formatType, size,
dateUploaded, and rentries fields for every data package on the
Arctic Data Center large data storage filesystem as a data.frame.
This can then be combined with object counts from Metacat and used
to extract various metrics and plots.
}
\details{
The function works by running local bash commands on the datateam host, and
will not run on other hosts
}
