% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_objects.R
\name{query_objects}
\alias{query_objects}
\title{Return key information on all objects on the Arctic Data Center}
\usage{
query_objects(n = 1e+06)
}
\arguments{
\item{n}{(integer) Number of rows to return}
}
\value{
(data.frame) Result of the SOLR query
}
\description{
This function returns the identifier, origin, formatType, size,
dateUploaded, and obsoletedBy fields for every object on the
Arctic Data Center as a data.frame. This can then be used to extract
various metrics and plots.
}
