% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_support_interactions.R
\name{update_annual_tix}
\alias{update_annual_tix}
\title{Update text file of annual ticket events}
\usage{
update_annual_tix(year, path = paste0(getwd(), "/ticket_list.csv"))
}
\arguments{
\item{year}{Year to update}

\item{path}{path of ticket list to read from}
}
\description{
Update text file of annual ticket events
}
