% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_support_interactions.R
\name{update_ticket_list}
\alias{update_ticket_list}
\title{Update text file of all tickets}
\usage{
update_ticket_list(path = paste0(getwd(), "/ticket_list.csv"))
}
\arguments{
\item{path}{a path to write the ticket list to}
}
\description{
Update text file of all tickets
}
