% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative_metric.R
\name{plot_cumulative_metric}
\alias{plot_cumulative_metric}
\title{Plot cumulative metric}
\usage{
plot_cumulative_metric(
  objects,
  metric,
  from = as.Date("2009-01-01"),
  to = as.Date(Sys.Date()),
  ...
)
}
\arguments{
\item{objects}{(data.frame) Table obtained from \code{query_objects}}

\item{metric}{(char) one of "files" or "datasets"}

\item{from}{Start date of plot (chatacter or POSIXct)}

\item{to}{End date of plot (character of POSIXct)}

\item{...}{additional arguments to \code{plot_theme_adc}}
}
\value{
Plot of total datasets
}
\description{
Plot cumulative metric
}
