% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_version_chains.R
\name{query_version_chains}
\alias{query_version_chains}
\title{Query for version chains of all metadata objects on the Arctic Data Center}
\usage{
query_version_chains(objects, cache_tolerance = 14)
}
\arguments{
\item{objects}{(data.frame) Result from \code{query_objects}}

\item{cache_tolerance}{(integer) Maximum number of days tolerable for age of
cached results. If the cached results are older than the cache tolerance, a
new dataset will be cached and the old cache deleted. To force a cache
refresh, set tolerance to zero.}
}
\value{
(data.frame) Metadata objects data.frame with series identifier
}
\description{
This function takes a data frame of results from \code{query_objects}, filters for
metadata objects only, then assigns an arbitary series identifier to each
object.
}
