% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_support_interactions.R
\name{update_annual_tix}
\alias{update_annual_tix}
\title{Update text file of annual ticket events}
\usage{
update_annual_tix(year)
}
\arguments{
\item{Year}{Year to update}
}
\description{
Update text file of annual ticket events
}
